feat(linkedin): merge live + takeout into one connector
Fold linkedin.takeout into the linkedin connector (key="linkedin",
v3.0.0) so a single connection dedups people met live and people in the
CSV export on the shared linkedin_slug/email identity.

- One connector spans BOTH sources: live Chrome-extension feeds
  (home_feed, company_updates, jobs → posts/job_posting) and local Data
  Export CSV feeds (connections, messages, invitations, applied_jobs,
  profile, companies, learning, events, endorsements, media).
- sync(ctx) dispatches by feedKey: live feeds → syncLive (extension),
  file feeds → the CSV readers via result().
- Resolve the jobs key clash: takeout's own applied-jobs export becomes
  'applied_jobs' (checkpoint last_applied_jobs_timestamp); live company
  hiring listings keep 'jobs'.
- Identity attributions preserved verbatim on connections (slug+email,
  neither primary — tolerates LinkedIn slug changes) and messages
  (sender slug). Merged checkpoint unions live + takeout cursors.
- Consolidate lobu.config to one linkedin connection (slug
  linkedin-takeout-buremba → existing buremba conn 410, 2544 events).
- Delete linkedin-takeout.connector.ts; move its identity/CSV tests into
  the linkedin connector test (29 pass).

Verified live against the real export: connections 2544 (all
slug-keyed), messages 4599/5000 slug-carrying, applied_jobs 2, all
NUL-free checkpoints.
